chore: organise directories
