+++ src/hooks/chat/useMessageReactions.ts
@@ -3,7 +3,7 @@ import type { Dispatch, SetStateAction } from 'react'
 import { addToast } from '@heroui/react'
 import { useCallback, useRef, useState } from 'react'
 
-import { useSupabase } from '@/components/contexts/SupabaseContext'
+import { useSupabase } from '@/components/chat/SupabaseContext'
 import { logEvent } from '@/utils/tasks'
 
 export interface MessageReaction {
